+++ app/src/main/java/com/example/logifitappp/data/dao/commons/WearableRawActivityDao.kt
@@ -13,7 +13,7 @@ abstract class WearableRawActivityDao<T: WearableRawActivityModel>(private val t
     @Upsert
     abstract fun store(vararg activities: T)
 
-    fun getRawActivitiesBetween(from: Int, to: Int, wearableId: Int): List<T> {
+    fun getRawActivitiesBetween(from: Long, to: Long, wearableId: Int): List<T> {
         return getRawActivitiesBetween(
             SimpleSQLiteQuery("SELECT * FROM $tableName WHERE timestamp >= $from AND timestamp <= $to AND wearable_id = $wearableId ORDER BY timestamp ASC")
         )
